pyproject.toml
[tool.poetry]
name = "lms-project"
version = "0.1.0"
description = ""
authors = ["alersandr <zasbox@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
django = "^5.2.6"
psycopg2-binary = "^2.9.10"
pillow = "^11.3.0"
dotenv = "^0.9.9"
djangorestframework = "^3.16.1"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
